add tests, divide code in different files
